+++ src/constants.ts
@@ -1,20 +1,2 @@
-export const VALID_CLIENTS = [
-	"claude",
-	"cline",
-	"windsurf",
-	"roocode",
-	"witsy",
-	"enconvo",
-	"cursor",
-	"vscode",
-	"vscode-insiders",
-	"boltai",
-	"amazon-bedrock",
-	"amazonq",
-	"librechat",
-	"gemini-cli",
-] as const
-export type ValidClient = (typeof VALID_CLIENTS)[number]
-
 export const REGISTRY_ENDPOINT = process.env.REGISTRY_ENDPOINT
 export const ANALYTICS_ENDPOINT = process.env.ANALYTICS_ENDPOINT